Add permission exception if try to delete the deviceId of other user

+++ src/security/application/security.service.ts
@@ -2,14 +2,19 @@ import {
   authCommandRepository,
   authQueryRepository,
 } from "../../auth/repositories";
-import { NotFoundException } from "../../common/exceptions";
+import {
+  NotFoundException,
+  PermissionException,
+} from "../../common/exceptions";
 
 export const securityService = {
-  async deleteDevice(deviceId: string): Promise<void> {
+  async deleteDevice(userId: string, deviceId: string): Promise<void> {
     const sessionToDelete =
       await authQueryRepository.getSessionByDeviceId(deviceId);
 
     if (!sessionToDelete) throw new NotFoundException();
+
+    if (sessionToDelete.userId !== userId) throw new PermissionException();
 
     await authCommandRepository.deleteSessionByDeviceId(deviceId);
   },
